Rewrite LCCH
It was previously fairly closely based on Lunar's proprietary implementation.
I did try to make it a bit more original previously, but that was with little knowledge of how it worked so it ended up looking a bit too similar to Lunar's code.
Lunar's implementation is actually broken, but this commit fixes that!