Remove launcher (#123)
* Bye launcher :cry:!

This launcher all began on the 31st of October, 2020. It was made in
Java, and only capable of launching the vanilla game.

Soon afterwards, I tried developing an actual client to launch. I didn't
want to use MCP, because I prefered mixin. I'd already made a standalone
mixin server using LaunchWrapper and looking at
https://github.com/CodeCrafter47/Rainbow/ amongst other projects - for
1.16.

However, I couldn't get either Legacy Fabric or ForgeGradle to work
nicely, neither could I get the LaunchWrapper service to work for some
reason (plus I didn't have a Discord account). I somehow ended up doing
asm which wasn't nice. I tried using some build-time stuff to remap
string constants containing mappings. The first feature of this client I
tried to implement was something to make it so that pressing F3+C *only*
cleared the chat rather than bringing up the whole overlay too. This was
hard!

Eventually I gave up and after explaining to my friend that it was
causing me to (metaphorically) tear my hair out, I finally opted for
MCP. Later on I managed to get MCP working with OptiFine and distributed
the launcher with a crude autoupdater to this friend.

One day, I spammed !polycatgen commands on the PolyMC Discord (before
all of the drama happened - well at least the most recent one) and
nearly got banned or muted (I can't remember). This is unrelated, but
funny! On the same day (I can't verify because I'm now banned!), I
started working on an electron launcher but it didn't go anywhere.

When I saw my friend had migrated, I tried to add a feature to support
this. Since the launcher code was made before Microsoft accounts to play
the game were a thing, adding it was a lot of effort. I couldn't work
out how to register an application so used a webview (mistake).
MiniDigger had a good JavaFX example but this project was swing so it
needed to be embedded. This didn't go well and the Linux JDK doesn't
have JavaFX bundled so eventually I returned to the Electron launcher.
This is the launcher this commit finally rids the world of (if merged)!

p.s. my hair has been growing without being cut this whole time which is
cool

* Hopefully fix build

* Add slightly important files

* Readd README

* Update outdated README

* Update HEADER location

* Add launcher note

* Use blossom for version replacement

* Hopefully fix build

* Hopefully fix build (deja vu)

* Hopefully fix build (more déjà vu)